refactor : Mapper통일에 따른 테스트 코드 통일
